% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_run.r
\name{seurat_run}
\alias{seurat_run}
\title{Run Seurat for a simulated data set}
\usage{
seurat_run(sim, ntop, out_seu = NULL, res = NULL, dims.use = NULL,
  test.de)
}
\arguments{
\item{sim}{A Splatter sim object with simulated cellular groups or paths.}

\item{ntop}{The absolute number of top ranked markers per cluster. 
It can be set according to the number of sim.markers (based on the spec parameter in the markers_by_specificity).}

\item{out_seu}{The Seurat output of seurat_run. If provided, it recompute the set of group markers, 
without rerun Seurat. It is helpful when you need to change the ntop parameter.}

\item{res}{The value of resolution in the Seurat FindClusters function. Default is set to 0.8.}

\item{dims.use}{The vector with the dimensions to use in the Seurat FindClusters function. Default is set to 1:2.}

\item{test.de}{The test used to run the Seurat FindAllMarkers function.}
}
\value{
A list with clusters, the output of the Seurat FindAllMarkers function, the set of genes per cluster and the running time. 
If out_seu is provided, only the set of genes per cluster is returned.
}
\description{
This function runs the Seurat package for a data set simulated by using Splatter
}
\examples{

}
